+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### v2.0.2 - Paging Feature in asset list and Cosmetic Changes (Current)
 - **[MFW-11](https://github.com/alperozdamar/my-finance-watcher/pull/24)** - Showing dollar/TL Exchance Rate (Online)
+- **[MFW-12]
 
 ### v2.0.1 - Displaying the running version in the footer 
 - **[MFW-10](https://github.com/alperozdamar/my-finance-watcher/pull/22)** - Showing dollar/TL Exchance Rate (Online)