+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 ## 📦 Releases / Changelog
 
-### v2.0.2 - Paging Feature in asset list and Cosmetic Changes (Current)
+### v2.0.3 - Available RSU is added (Current)
+- **[MFW-13](https://github.com/alperozdamar/my-finance-watcher/pull/26)** - Adding column name to Available RSU
+
+### v2.0.2 - Paging Feature in asset list and Cosmetic Changes
 - **[MFW-12](https://github.com/alperozdamar/my-finance-watcher/pull/25)** - Changing column name to retSerda
 - **[MFW-11](https://github.com/alperozdamar/my-finance-watcher/pull/24)** - Paging Feature in asset list is added